+++ NEWS.md
@@ -1,9 +1,16 @@
 # Changelog
 
-## 2.31.2
+## 2.31.3
 
 Released 2016-mm-dd
 
+## 2.31.2
+
+Released 2016-04-04
+
+Bug fixes:
+ - Overviews integration for named layers #400
+ - Support wrapped queries in named layers #405
 
 ## 2.31.1